backend/pyproject.toml
[project]
name = "versefinder-api"
version = "0.1.0"
description = "FastAPI backend for VerseFinder AI - semantic Bible verse search"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "faiss-cpu>=1.9.0",
    "numpy>=2.0.0",
    "openai>=1.50.0",
    "bm25s>=0.2.0",
    "PyStemmer>=2.2.0",
    "pydantic-settings>=2.0.0",
    "google-generativeai>=0.8.0",  # Gemini LLM provider
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",  # For testing FastAPI
]

# Run commands:
# uv run uvicorn app.main:app --reload
# uv run pytest
